+++ examples/ShoppingAds/AddPerformanceMaxRetailCampaign.php
@@ -25,49 +25,49 @@ use Google\Ads\GoogleAds\Examples\Utils\ArgumentNames;
 use Google\Ads\GoogleAds\Examples\Utils\ArgumentParser;
 use Google\Ads\GoogleAds\Examples\Utils\Helper;
 use Google\Ads\GoogleAds\Lib\OAuth2TokenBuilder;
-use Google\Ads\GoogleAds\Lib\V16\GoogleAdsClient;
-use Google\Ads\GoogleAds\Lib\V16\GoogleAdsClientBuilder;
-use Google\Ads\GoogleAds\Lib\V16\GoogleAdsException;
+use Google\Ads\GoogleAds\Lib\V17\GoogleAdsClient;
+use Google\Ads\GoogleAds\Lib\V17\GoogleAdsClientBuilder;
+use Google\Ads\GoogleAds\Lib\V17\GoogleAdsException;
 use Google\Ads\GoogleAds\Util\FieldMasks;
-use Google\Ads\GoogleAds\Util\V16\ResourceNames;
-use Google\Ads\GoogleAds\V16\Common\ImageAsset;
-use Google\Ads\GoogleAds\V16\Common\LanguageInfo;
-use Google\Ads\GoogleAds\V16\Common\LocationInfo;
-use Google\Ads\GoogleAds\V16\Common\MaximizeConversionValue;
-use Google\Ads\GoogleAds\V16\Common\TextAsset;
-use Google\Ads\GoogleAds\V16\Enums\AdvertisingChannelTypeEnum\AdvertisingChannelType;
-use Google\Ads\GoogleAds\V16\Enums\AssetFieldTypeEnum\AssetFieldType;
-use Google\Ads\GoogleAds\V16\Enums\AssetGroupStatusEnum\AssetGroupStatus;
-use Google\Ads\GoogleAds\V16\Enums\BudgetDeliveryMethodEnum\BudgetDeliveryMethod;
-use Google\Ads\GoogleAds\V16\Enums\CampaignStatusEnum\CampaignStatus;
-use Google\Ads\GoogleAds\V16\Enums\ConversionActionCategoryEnum\ConversionActionCategory;
-use Google\Ads\GoogleAds\V16\Enums\ConversionOriginEnum\ConversionOrigin;
-use Google\Ads\GoogleAds\V16\Enums\ListingGroupFilterListingSourceEnum\ListingGroupFilterListingSource;
-use Google\Ads\GoogleAds\V16\Enums\ListingGroupFilterTypeEnum\ListingGroupFilterType;
-use Google\Ads\GoogleAds\V16\Errors\GoogleAdsError;
-use Google\Ads\GoogleAds\V16\Resources\Asset;
-use Google\Ads\GoogleAds\V16\Resources\AssetGroup;
-use Google\Ads\GoogleAds\V16\Resources\AssetGroupAsset;
-use Google\Ads\GoogleAds\V16\Resources\AssetGroupListingGroupFilter;
-use Google\Ads\GoogleAds\V16\Resources\Campaign;
-use Google\Ads\GoogleAds\V16\Resources\Campaign\ShoppingSetting;
-use Google\Ads\GoogleAds\V16\Resources\CampaignBudget;
-use Google\Ads\GoogleAds\V16\Resources\CampaignConversionGoal;
-use Google\Ads\GoogleAds\V16\Resources\CampaignCriterion;
-use Google\Ads\GoogleAds\V16\Services\AssetGroupAssetOperation;
-use Google\Ads\GoogleAds\V16\Services\AssetGroupListingGroupFilterOperation;
-use Google\Ads\GoogleAds\V16\Services\AssetGroupOperation;
-use Google\Ads\GoogleAds\V16\Services\AssetOperation;
-use Google\Ads\GoogleAds\V16\Services\CampaignBudgetOperation;
-use Google\Ads\GoogleAds\V16\Services\CampaignConversionGoalOperation;
-use Google\Ads\GoogleAds\V16\Services\CampaignCriterionOperation;
-use Google\Ads\GoogleAds\V16\Services\CampaignOperation;
-use Google\Ads\GoogleAds\V16\Services\GoogleAdsRow;
-use Google\Ads\GoogleAds\V16\Services\MutateGoogleAdsRequest;
-use Google\Ads\GoogleAds\V16\Services\MutateGoogleAdsResponse;
-use Google\Ads\GoogleAds\V16\Services\MutateOperation;
-use Google\Ads\GoogleAds\V16\Services\MutateOperationResponse;
-use Google\Ads\GoogleAds\V16\Services\SearchGoogleAdsRequest;
+use Google\Ads\GoogleAds\Util\V17\ResourceNames;
+use Google\Ads\GoogleAds\V17\Common\ImageAsset;
+use Google\Ads\GoogleAds\V17\Common\LanguageInfo;
+use Google\Ads\GoogleAds\V17\Common\LocationInfo;
+use Google\Ads\GoogleAds\V17\Common\MaximizeConversionValue;
+use Google\Ads\GoogleAds\V17\Common\TextAsset;
+use Google\Ads\GoogleAds\V17\Enums\AdvertisingChannelTypeEnum\AdvertisingChannelType;
+use Google\Ads\GoogleAds\V17\Enums\AssetFieldTypeEnum\AssetFieldType;
+use Google\Ads\GoogleAds\V17\Enums\AssetGroupStatusEnum\AssetGroupStatus;
+use Google\Ads\GoogleAds\V17\Enums\BudgetDeliveryMethodEnum\BudgetDeliveryMethod;
+use Google\Ads\GoogleAds\V17\Enums\CampaignStatusEnum\CampaignStatus;
+use Google\Ads\GoogleAds\V17\Enums\ConversionActionCategoryEnum\ConversionActionCategory;
+use Google\Ads\GoogleAds\V17\Enums\ConversionOriginEnum\ConversionOrigin;
+use Google\Ads\GoogleAds\V17\Enums\ListingGroupFilterListingSourceEnum\ListingGroupFilterListingSource;
+use Google\Ads\GoogleAds\V17\Enums\ListingGroupFilterTypeEnum\ListingGroupFilterType;
+use Google\Ads\GoogleAds\V17\Errors\GoogleAdsError;
+use Google\Ads\GoogleAds\V17\Resources\Asset;
+use Google\Ads\GoogleAds\V17\Resources\AssetGroup;
+use Google\Ads\GoogleAds\V17\Resources\AssetGroupAsset;
+use Google\Ads\GoogleAds\V17\Resources\AssetGroupListingGroupFilter;
+use Google\Ads\GoogleAds\V17\Resources\Campaign;
+use Google\Ads\GoogleAds\V17\Resources\Campaign\ShoppingSetting;
+use Google\Ads\GoogleAds\V17\Resources\CampaignBudget;
+use Google\Ads\GoogleAds\V17\Resources\CampaignConversionGoal;
+use Google\Ads\GoogleAds\V17\Resources\CampaignCriterion;
+use Google\Ads\GoogleAds\V17\Services\AssetGroupAssetOperation;
+use Google\Ads\GoogleAds\V17\Services\AssetGroupListingGroupFilterOperation;
+use Google\Ads\GoogleAds\V17\Services\AssetGroupOperation;
+use Google\Ads\GoogleAds\V17\Services\AssetOperation;
+use Google\Ads\GoogleAds\V17\Services\CampaignBudgetOperation;
+use Google\Ads\GoogleAds\V17\Services\CampaignConversionGoalOperation;
+use Google\Ads\GoogleAds\V17\Services\CampaignCriterionOperation;
+use Google\Ads\GoogleAds\V17\Services\CampaignOperation;
+use Google\Ads\GoogleAds\V17\Services\GoogleAdsRow;
+use Google\Ads\GoogleAds\V17\Services\MutateGoogleAdsRequest;
+use Google\Ads\GoogleAds\V17\Services\MutateGoogleAdsResponse;
+use Google\Ads\GoogleAds\V17\Services\MutateOperation;
+use Google\Ads\GoogleAds\V17\Services\MutateOperationResponse;
+use Google\Ads\GoogleAds\V17\Services\SearchGoogleAdsRequest;
 use Google\ApiCore\ApiException;
 use Google\ApiCore\Serializer;
 
@@ -117,8 +117,6 @@ class AddPerformanceMaxRetailCampaign
     // temporary IDs because they are referenced only once.
     /** @var int the negative temporary ID used in bulk mutates. */
     private static $nextTempId = self::ASSET_GROUP_TEMPORARY_ID - 1;
-
-    private const PAGE_SIZE = 10000;
 
     public static function main()
     {
@@ -828,9 +826,8 @@ class AddPerformanceMaxRetailCampaign
             'FROM customer_conversion_goal';
         // The number of conversion goals is typically less than 50 so we use a search request
         // instead of search stream.
-        $response = $googleAdsServiceClient->search(
-            SearchGoogleAdsRequest::build($customerId, $query)->setPageSize(self::PAGE_SIZE)
-        );
+        $response =
+            $googleAdsServiceClient->search(SearchGoogleAdsRequest::build($customerId, $query));
 
         // Iterates over all rows in all pages and builds the list of conversion goals.
         foreach ($response->iterateAllElements() as $googleAdsRow) {